requirements/requirements.txt
appconfigpy>=0.4.0,<2
click>=6.2,<8.0
colorama>=0.3.7,<1.0.0
Logbook>=0.12.3,<2.0.0
DateTimeRange>=0.6.1,<2
msgfy>=0.0.6,<2
path>=13.0.0,<14
pathvalidate>=1.1.0,<2
PyGithub>=1.43.7,<2.0.0
retryrequests>=0.0.2,<1
typepy>=0.6.4,<2
typing-extensions>=3.7.4
